feat: simplify chat to new-chat and chat-session components
