--- frontend/src/components/conversation-view/new-conversation.tsx
@@ -1,57 +1,0 @@
-import { useApplicationContext } from "@/lib/use-application-context.ts";
-import { useNavigate } from "react-router-dom";
-import { MessageComposer } from "@/components/conversation-view/message-composer.tsx";
-import { ApiClient } from "@/lib/api.ts";
-import { authenticationProviderInstance } from "@/lib/authentication-provider.ts";
-import { useState, useEffect } from "react";
-import { AgentDetails } from "@/lib/types";
-
-import type { OnPendingMessage } from "@/components/conversation-view/chat";
-
-export interface NewConversationProps {
-  agentId: number;
-  onPendingMessage: OnPendingMessage;
-}
-
-const api = new ApiClient(authenticationProviderInstance);
-
-export function NewConversation({ agentId, onPendingMessage }: NewConversationProps) {
-  const [agent, setAgent] = useState<AgentDetails>();
-
-  useEffect(() => {
-    const loadAgentDetails = async () => {
-      try {
-        const agentDetails = await api.agents().getAgentById(agentId);
-        setAgent(agentDetails);
-      } catch (error) {
-        console.error('Failed to load agent details:', error);
-      }
-    };
-
-    loadAgentDetails();
-  }, [agentId]);
-
-  const onSubmit = async (message: string) => {
-    const newConversationId = await api.conversations().createConversation(agentId);
-    onPendingMessage(message, newConversationId);
-  }
-
-  return (
-    <div className="flex flex-col h-full px-4 pt-4">
-      <div className="grow flex items-start justify-center pt-16">
-        {agent?.description && (
-          <div className="text-center space-y-4 max-w-2xl px-6">
-            <div className="bg-background rounded-lg p-6">
-              <p className="text-gray-500 text-base leading-relaxed">
-                {agent.description}
-              </p>
-            </div>
-          </div>
-        )}
-      </div>
-      <div className="bottom-0 sticky flex-shrink-0 bg-white pb-4">
-        <MessageComposer onSubmit={onSubmit} isLoading={false}/>
-      </div>
-    </div>
-  )
-}
